requirements.txt
# === Core ===
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
filelock==3.16.1
idna==3.7
requests==2.32.3
urllib3==2.2.3
python-dotenv==1.0.1
python-decouple==3.8
gunicorn
packaging==24.1
pytz==2024.2
typing_extensions==4.12.2

# === Django + REST ===
Django==4.2.16
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
sqlparse==0.5.2

# === LangChain stack (stable combo) ===
langchain==0.3.7
langchain-community==0.3.6
langchain-core==0.3.16
langchain-text-splitters==0.3.2
SQLAlchemy==2.0.35
langsmith==0.1.136

# === NLP (spaCy with CORRECT blis version) ===
spacy==3.7.2
blis==0.7.11
cymem==2.0.13
preshed==3.0.12
thinc==8.2.3
srsly==2.5.2
wasabi==1.1.3
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

# === Data + Utils (Python 3.10 compatible) ===
pandas==2.1.4
numpy==1.26.4
pydantic==2.9.2
pydantic-core==2.23.4
PyYAML==6.0.2
openpyxl==3.1.5
python-docx==1.1.2
fpdf2==2.8.2
reportlab==4.2.5

# === Optional lightweight tools ===
beautifulsoup4==4.12.3
html5lib==1.1
fake-useragent==2.2.0
pytube==15.0.0
youtube-transcript-api==0.6.3
yt-dlp==2025.1.15